service/requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.1
certifi==2019.11.28
cfgv==3.1.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.4
django-filter==2.2.0
django-mysql==3.3.0
djangorestframework==3.11.0
drf-yasg==1.17.1
filelock==3.0.12
identify==1.4.11
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
mysqlclient==1.4.6
nodeenv==1.3.5
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pre-commit==2.1.1
psycopg2-binary==2.8.4
PyJWT==1.7.1
pyparsing==2.4.6
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
virtualenv==20.0.10
